+++ src/OpenTelemetry.Instrumentation.EntityFrameworkCore/CHANGELOG.md
@@ -12,6 +12,9 @@
   conventions when opted into using the `OTEL_SEMCONV_STABILITY_OPT_IN` environment
   variable.
   ([#3011](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3011))
+* Add `db.query.parameter.<key>` attribute(s) to query spans if opted into using
+  the `SetDbQueryParameters` option.
+  ([#3015](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3015))
 
 ## 1.12.0-beta.2